requirements.txt
datasets==2.15.0
jiwer==3.0.3
numpy==1.26.2
openai_whisper==20231117
opencc_python_reimplemented==0.1.7
pandas==2.1.3
torch==2.0.0
tqdm==4.66.1
transformers==4.35.2
lingua-language-detector==2.0.2
librosa==0.10.1
soundfile==0.12.1
